+++ src/PrintOrder/PrintOrder.js
@@ -3,16 +3,20 @@ import PropTypes from 'prop-types';
 import { useReactToPrint } from 'react-to-print';
 import { useIntl } from 'react-intl';
 
-import { stripesConnect } from '@folio/stripes/core';
+import { stripesConnect, useOkapiKy } from '@folio/stripes/core';
 
 import { exportManifest, getExportData } from '../common/ExportSettingsModal/utils';
 
 import PrintContent from './PrintContent';
 import { hydrateOrderToPrint } from './hydrateOrderToPrint';
-import { getPrintPageStyles } from './utils';
+import {
+  getOrderPrintData,
+  getPrintPageStyles,
+} from './utils';
 
 export const PrintOrderComponent = ({ mutator, order, orderLine, onCancel }) => {
   const intl = useIntl();
+  const ky = useOkapiKy();
 
   const [printableOrder, setPrintableOrder] = useState();
 
@@ -27,11 +31,14 @@ export const PrintOrderComponent = ({ mutator, order, orderLine, onCancel }) => 
     (async () => {
       const { compositePoLines } = order;
       const linesToPrint = orderLine ? [orderLine] : compositePoLines;
+      const printData = compositePoLines?.length
+        ? { lines: await getExportData(mutator, linesToPrint, [order], intl) }
+        : await getOrderPrintData(ky, order);
 
       setPrintableOrder(hydrateOrderToPrint({
         order: {
           ...order,
-          lines: await getExportData(mutator, linesToPrint, [order], intl),
+          ...printData,
         },
       }));