+++ src/PrintOrder/hydrateOrderToPrint.js
@@ -4,10 +4,14 @@ export function hydrateOrderToPrint({ order }) {
   }
 
   let totals = {};
-  const billToAddress = order.lines?.[0].billToRecord?.address;
-  const shipToAddress = order.lines?.[0].shipToRecord?.address;
-  const vendorPrimaryAddress = order.lines?.[0].vendorRecord?.addresses?.find(({ isPrimary }) => isPrimary);
-  const vendorPrimaryPhone = order.lines?.[0].vendorRecord?.phoneNumbers?.find(({ isPrimary }) => isPrimary);
+
+  const billToAddress = order.billToRecord?.address || order.lines?.[0]?.billToRecord?.address;
+  const shipToAddress = order?.shipToRecord?.address || order?.lines?.[0]?.shipToRecord?.address;
+  const vendor = order.vendorRecord || order.lines?.[0]?.vendorRecord;
+  const vendorPrimaryAddress = (order.vendorRecord?.addresses || order.lines?.[0]?.vendorRecord?.addresses)
+    ?.find(({ isPrimary }) => isPrimary);
+  const vendorPrimaryPhone = (order.vendorRecord?.phoneNumbers || order.lines?.[0]?.vendorRecord?.phoneNumbers)
+    ?.find(({ isPrimary }) => isPrimary);
 
   if (order.lines?.length === 1) {
     const [{ poLineEstimatedPrice, quantityPhysical = 0, quantityElectronic = 0 }] = order.lines;
@@ -25,6 +29,6 @@ export function hydrateOrderToPrint({ order }) {
     shipToAddress,
     vendorPrimaryAddress,
     vendorPrimaryPhone,
-    vendor: order.lines?.[0].vendorRecord,
+    vendor,
   };
 }