+++ src/PrintOrder/utils.js
@@ -1,5 +1,19 @@
 import { pick } from 'lodash';
 
+import {
+  VENDORS_API,
+  CONFIG_API,
+} from '@folio/stripes-acq-components';
+
+import {
+  CONFIG_ADDRESSES,
+  MODULE_TENANT,
+} from '../components/Utils/const';
+import {
+  getAddresses,
+  getRecordMap,
+} from '../common/utils';
+
 export const getPrintPageStyles = () => `
   @page {
     size: A4 landscape;
@@ -30,3 +44,23 @@ export const buildAddressString = (address = {}) => (
     .filter(Boolean)
     .join(', ')
 );
+
+export const getOrderPrintData = async (ky, order = {}) => {
+  const vendor = await ky.get(`${VENDORS_API}/${order.vendor}`).json().catch(() => ({}));
+  const addressIds = [...new Set([order.billTo, order.shipTo])].filter(Boolean);
+  const subQuery = addressIds.map(id => `id==${id}`).join(' or ');
+  const searchParams = {
+    query: `(module=${MODULE_TENANT} and configName=${CONFIG_ADDRESSES} and (${subQuery}))`,
+  };
+  const addressesResp = addressIds.length
+    ? await ky.get(CONFIG_API, { searchParams }).json().catch(() => ({}))
+    : {};
+  const addresses = getAddresses(addressesResp?.configs || []);
+  const addressMap = getRecordMap(addresses);
+
+  return ({
+    vendorRecord: vendor,
+    billToRecord: addressMap[order.billTo],
+    shipToRecord: addressMap[order.shipTo],
+  });
+};