+++ src/common/utils/getRecordMap.js
@@ -1,0 +1,7 @@
+export const getRecordMap = (records) => (
+  records.reduce((acc, record) => {
+    acc[record.id] = record;
+
+    return acc;
+  }, {})
+);